Bootstrap of ergmito

Bootstrap of ergmito

ergmito_boot(x, ..., R, ncpus = 1L, cl = NULL)

Arguments

x
Either a formula or an object of class ergmito.

...
Additional arguments passed to the method.

R
Integer. Number of replicates

ncpus
Integer Number of CPUs to use. Only recommended if ergmito was not compiled with OpenMP (otherwise it will be slower).

cl
An object of class cluster (see makePSOCKcluster)

Value

An object of class ergmito_boot and ergmito. This adds three elements to the ergmito object:

R The number of replicates.

sample A vector of length R with the cases used in each replicate.

dist The distribution of fitted parameters.

nvalid the number of cases used for computing the covar.

timer_boot records the time the whole process took.

Details

The resulting sample of parameters estimates is then used to compute the variance-covariance matrix of the model. Cases in which Inf/NaN/NA values were returned are excluded from the calculation.

Examples

data(fivenets)
set.seed(123)
ans0 <- ergmito(fivenets ~ edges + ttriad)

#> Warning: The observed statistics (target.statistics) are near or at the boundary of its support, i.e. the Maximum Likelihood Estimates maynot exist or be hard to be estimated. In particular, the statistic(s) "edges", "ttriple".

ans1 <- suppressWarnings(ergmito_boot(ans0, R = 100))

#> 
#> |0%                 |25%                |50%                |75%           100%|
#> --------------------------------------------------------------------------------
#> ////////////////////////////////////////////////////////////////////////////////

ans2 <- suppressWarnings(ergmito_boot(fivenets ~ edges + ttriad, R = 100))

#> 
#> |0%                 |25%                |50%                |75%           100%|
#> --------------------------------------------------------------------------------
#> ////////////////////////////////////////////////////////////////////////////////

# Checking the differences
summary(ans0)

#> 
#> ERGMito estimates (MLE)
#> This model includes 5 networks.
#> 
#> formula:
#>   fivenets ~ edges + ttriad
#> <environment: 0xda31a30>
#> 
#>         Estimate Std. Error z value Pr(>|z|)  
#> edges   -0.89666    0.39389 -2.2765  0.02282 *
#> ttriple  0.26828    0.36290  0.7393  0.45973  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> AIC: 79.92984    BIC: 84.11852    (Smaller is better.) 

summary(ans1)

#> 
#> ERGMito estimates (MLE)
#> This model includes 5 networks.
#> 
#> (bootstrapped model with  100  replicates.)
#> 
#> formula:
#>   fivenets ~ edges + ttriad
#> <environment: 0xda31a30>
#> 
#>         Estimate Std. Error z value Pr(>|z|)   
#> edges   -0.89666    0.32632 -2.7478 0.005999 **
#> ttriple  0.26828    0.23361  1.1484 0.250792   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> AIC: 79.92984    BIC: 84.11852    (Smaller is better.) 

summary(ans2)

#> 
#> ERGMito estimates (MLE)
#> This model includes 5 networks.
#> 
#> (bootstrapped model with  100  replicates.)
#> 
#> formula:
#>   fivenets ~ edges + ttriad
#> <environment: 0xda31a30>
#> 
#>         Estimate Std. Error z value Pr(>|z|)  
#> edges   -0.89666    0.39037 -2.2970  0.02162 *
#> ttriple  0.26828    0.23119  1.1605  0.24586  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> AIC: 79.92984    BIC: 84.11852    (Smaller is better.)