Extract function to be used with the texreg package.

To be used with the texreg package. This function can be used to generate nice looking tables of ERGMitos estimates.

extract.ergmito(
  model,
  include.aic = TRUE,
  include.bic = TRUE,
  include.loglik = TRUE,
  include.nnets = TRUE,
  include.convergence = TRUE,
  ...
)

Arguments

model
An object of class ergmito.

include.aic, include.bic, include.loglik
See texreg::extract.

include.nnets
Logical. When true, it adds the Number of networks used to the list of gof statistics. This can be useful when running pooled models.

include.convergence
Logical. When true it, adds the convergence value of the stats::optim function (0 means convergence).

...
Further arguments passed to the base::summary() of ergmito.

Examples

library(texreg)

#> Version:  1.36.23
#> Date:     2017-03-03
#> Author:   Philip Leifeld (University of Glasgow)
#> 
#> Please cite the JSS article in your publications -- see citation("texreg").

data(fivenets)
ans <- ergmito(fivenets ~ edges + nodematch("female"))
screenreg(ans)

#> 
#> ===========================
#>                   Model 1  
#> ---------------------------
#> edges              -1.70 **
#>                    (0.54)  
#> nodematch.female    1.59 * 
#>                    (0.64)  
#> ---------------------------
#> AIC                73.34   
#> BIC                77.53   
#> Log Likelihood    -34.67   
#> Num. networks       5      
#> Convergence         0      
#> ===========================
#> *** p < 0.001, ** p < 0.01, * p < 0.05