Estimation of ERGMs using Exact likelihood functions

As a difference from ergm::ergm, ergmito uses the exact log-likelihood function for fitting the model. This implies that all the 2^(n*(n-1)) graphs are generated for computing the normalizing constant of the ERGM model. This implies that models with up to 5 nodes are relatively simple to fit, since more than that can become infeasible.

ergmito(
  model,
  gattr_model = NULL,
  stats.weights = NULL,
  stats.statmat = NULL,
  optim.args = list(),
  init = NULL,
  use.grad = TRUE,
  target.stats = NULL,
  ntries = 1L,
  ncores = 1L,
  keep.stats = TRUE,
  ...
)

# S3 method for ergmito
print(x, ...)

# S3 method for ergmito
summary(object, ...)

# S3 method for ergmito_summary
print(x, ...)

# S3 method for ergmito
coef(object, ...)

# S3 method for ergmito
logLik(object, ...)

# S3 method for ergmito
nobs(object, ...)

# S3 method for ergmito
vcov(object, solver = NULL, ...)

# S3 method for ergmito
formula(x, ...)

Arguments

model
Model to estimate. See ergm::ergm. The only difference with ergm is that the LHS can be a list of networks.

gattr_model
A formula. Model specification for graph attributes. This is useful when using multiple networks.

stats.weights, stats.statmat
Lists as returned by ergm::ergm.allstats. When this is provided, the function does not call ergm.allstats, which can be useful in simulations.

optim.args
List. Passed to stats::optim.

init
A numeric vector. Sets the starting parameters for the optimization routine. Default is a vector of zeros.

use.grad
Logical. When TRUE passes the gradient function to optim. This is intended for testing only (internal use).

target.stats
A matrix of target statistics (see ergm::ergm).

ntries
Integer scalar. Number of tries to estimate the MLE (see details).

ncores
Integer scalar. Number of cores (threads) for parallel computation.

keep.stats
Logical scalar. When TRUE (the default), the matrices and vectors associated with the sufficient statistics will be returned. Otherwise the function discards them. This may be useful for saving memory space when estimating multiple models.

...
Further arguments passed to the method. In the case of ergmito, ... are passed to ergmito_formulae.

x, object
An object of class ergmito

solver
Function. Used to compute the inverse of the hessian matrix. When not null, the variance-covariance matrix is recomputed using that function. By default, ergmito uses MASS::ginv.

Value

An list of class ergmito:

call The program call.

coef Named vector. Parameter estimates.

iterations Integer. Number of times the log-likelihood was evaluated (see stats::optim).

mle.lik Numeric. Final value of the objective function.

null.lik Numeric. Final value of the objective function for the null model.

covar Square matrix of size length(coef). Variance-covariance matrix computed using the exact hessian as implemented in exact_hessian.

coef.init Named vector of length length(coef). Initial set of parameters used in the optimization.

formulae An object of class ergmito_loglik.

nobs Integer scalar. Number of networks in the model.

network Networks passed via model.

optim.out,optim.args Results from the optim call and arguments passed to it.

status,note Convergence code. See check_convergence

best_try Integer scalar. Index of the run with the highest log-likelihood value.

history Matrix of size ntries * (k + 1). History of the parameter estimates and the reached log-likelihood values.

timer Vector of times (for benchmarking). Each unit marks the starting point of the step.

MLE

Maximum Likelihood Estimates are obtained using the stats::optim function. The default method for maximization is BFGS using both the log-likelihood function and its corresponding gradient.

Another important factor to consider is the existence of the MLE estimates As shown in Handcock (2003), if the observed statistics are near the border if the support function (e.g. too many edges or almost none), then, even if the MLE estimates exists, the optimization function may not be able to reach the optima. Moreover, if the target (observed) statistics live in the boundary, then the MLE estimates do not exists. In general, this should not be an issue in the context of the pooled model, as the variability of observed statistics should be enough to avoid those situations.

The function ergmito will try to identify possible cases of non-existence, of the MLE, and if identified then try to re estimate the model parameters using larger values than the ones obtained, if the log-likelihood is greater, then it is assumed that the model is degenerate and the corresponding values will be replaced with either +Inf or -Inf. By default, this behavior is checked anytime that the absolute value of the estimates is greater than 5, or the sufficient statistics were flagged as potentially outside of the interior of the support (close to zero or to its max).

In the case of ntries, the optimization is repeated that number of times, each time perturbing the init parameter by adding a Normally distributed vector. The result which reaches the highest log-likelihood will be the one reported as parameter estimates. This feature is intended for testing only. Anecdotally, optim reaches the max in the first try.

See also

Examples

# Generating a small graph
set.seed(12)
n <- 4
net <- rbernoulli(n, p = .7)

model <- net ~ edges + mutual

library(ergm)

#> Loading required package: network

#> network: Classes for Relational Data
#> Version 1.16.0 created on 2019-11-30.
#> copyright (c) 2005, Carter T. Butts, University of California-Irvine
#>                     Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Martina Morris, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>  For citation information, type citation("network").
#>  Type help("network-package") to get started.

#> 
#> ergm: version 3.10.4, created on 2019-06-10
#> Copyright (c) 2019, Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Carter T. Butts, University of California -- Irvine
#>                     Steven M. Goodreau, University of Washington
#>                     Pavel N. Krivitsky, University of Wollongong
#>                     Martina Morris, University of Washington
#>                     with contributions from
#>                     Li Wang
#>                     Kirk Li, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>                     Chad Klumb
#> Based on "statnet" project software (statnet.org).
#> For license and citation information see statnet.org/attribution
#> or type citation("ergm").

#> NOTE: Versions before 3.6.1 had a bug in the implementation of the bd()
#> constriant which distorted the sampled distribution somewhat. In
#> addition, Sampson's Monks datasets had mislabeled vertices. See the
#> NEWS and the documentation for more details.

#> NOTE: Some common term arguments pertaining to vertex attribute and
#> level selection have changed in 3.10.0. See terms help for more
#> details. Use ‘options(ergm.term=list(version="3.9.4"))’ to use old
#> behavior.

ans_ergmito <- ergmito(model)
ans_ergm  <- ergm(model)

#> Starting maximum pseudolikelihood estimation (MPLE):

#> Evaluating the predictor and response matrix.

#> Maximizing the pseudolikelihood.

#> Finished MPLE.

#> Starting Monte Carlo maximum likelihood estimation (MCMLE):

#> Iteration 1 of at most 20:

#> Optimizing with step length 1.

#> The log-likelihood improved by 0.001321.

#> Step length converged once. Increasing MCMC sample size.

#> Iteration 2 of at most 20:

#> Optimizing with step length 1.

#> Error in solve.default(H, tol = 1e-20) : 
#>   Lapack routine dgesv: system is exactly singular: U[2,2] = 0

#> Warning: Approximate Hessian matrix is singular. Standard errors due to MCMC approximation of the likelihood cannot be evaluated. This is likely due to insufficient MCMC sample size or highly correlated model terms.

#> The log-likelihood improved by 0.005171.

#> Step length converged twice. Stopping.

#> Finished MCMLE.

#> Evaluating log-likelihood at the estimate. 

#> Using 20 bridges: 

#> 1 

#> 2 

#> 3 

#> 4 

#> 5 

#> 6 

#> 7 

#> 8 

#> 9 

#> 10 

#> 11 

#> 12 

#> 13 

#> 14 

#> 15 

#> 16 

#> 17 

#> 18 

#> 19 

#> 20 

#> .

#> This model was fit using MCMC.  To examine model diagnostics and check
#> for degeneracy, use the mcmc.diagnostics() function.

# The ergmito should have a larger value
ergm.exact(ans_ergmito$coef, model)

#>           [,1]
#> [1,] -6.239738

ergm.exact(ans_ergm$coef, model)

#>           [,1]
#> [1,] -6.239572

summary(ans_ergmito)

#> 
#> ERGMito estimates
#> note: optim converged, but the Hessian is not p.s.d.
#> 
#> formula:  net ~ edges + mutual 
#> 
#>        Estimate Std. Error z value Pr(>|z|)
#> edges    6.9736    26.7198  0.2610   0.7941
#> mutual  -6.2765    26.7448 -0.2347   0.8145
#> AIC: 16.47948    BIC: 17.44929    (Smaller is better.) 

summary(ans_ergm)

#> 
#> ==========================
#> Summary of model fit
#> ==========================
#> 
#> Formula:   net ~ edges + mutual
#> <environment: 0x266a8100>
#> 
#> Iterations:  2 out of 20 
#> 
#> Monte Carlo MLE Results:
#>        Estimate Std. Error MCMC % z value Pr(>|z|)
#> edges     19.18         NA     NA      NA       NA
#> mutual   -18.45         NA     NA      NA       NA
#> 
#>      Null Deviance: 16.64  on 12  degrees of freedom
#>  Residual Deviance: 12.39  on 10  degrees of freedom
#>  
#> AIC: 16.39    BIC: 17.36    (Smaller is better.) 

# Example 2: Estimating an ERGMito using data with know DGP parameters -----
data(fivenets)

model1 <- ergmito(fivenets ~ edges + nodematch("female"))

#> Warning: `set_attrs()` is deprecated as of rlang 0.3.0
#> This warning is displayed once per session.

summary(model1) # This data has know parameters equal to -2.0 and 2.0

#> 
#> ERGMito estimates
#> 
#> formula:  fivenets ~ edges + nodematch("female") 
#> 
#>                  Estimate Std. Error z value Pr(>|z|)   
#> edges            -1.70475    0.54356 -3.1363 0.001711 **
#> nodematch.female  1.58697    0.64305  2.4679 0.013592 * 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> AIC: 73.34109    BIC: 77.52978    (Smaller is better.) 

if (FALSE) {
# Example 3: Likelihood ratio test using the lmtest R package

library(lmtest)
data(fivenets)
model1 <- ergmito(fivenets ~ edges + nodematch("female"))
model2 <- ergmito(fivenets ~ edges + nodematch("female") + mutual)

lrtest(model1, model2)
# Likelihood ratio test
# 
# Model 1: fivenets ~ edges + nodematch("female") 
# Model 2: fivenets ~ edges + nodematch("female") + mutual
#   #Df  LogLik Df  Chisq Pr(>Chisq)
# 1   2 -34.671                     
# 2   3 -34.205 1 0.9312     0.3346

}