Processing formulas in ergmito

Analyze formula objects returning the matrices of weights and sufficient statistics to be used in the model together with the log-likelihood and gradient functions for joint models.

ergmito_formulae(model, gattr_model = NULL, target.stats = NULL,
  stats.weights = NULL, stats.statmat = NULL, env = parent.frame(),
  ...)

Arguments

model
A formula. The left-hand-side can be either a small network, or a list of networks.

gattr_model
A formula. Model especification for graph attributes. This is useful when using multiple networks.

target.stats
Observed statistics. If multiple networks, then a list, otherwise a named vector (see ergm::summary_formula).

stats.weights, stats.statmat
Lists of sufficient statistics and their respective weights.

env
Environment in which model should be evaluated.

...
Further arguments passed to ergm::ergm.allstats.

Value

A list of class ergmito_loglik.

loglik A function. The log-likelihood function.

grad A function. The gradient of the model.

stats.weights,stats.statmat two list of objects as returned by ergm::ergm.allstats.

model A formula. The model passed.

npars Integer. Number of parameters.

nnets Integer. Number of networks to estiamte.

vertex.attr Character vector. Vertex attributes used in the model.